conda-requirements.txt
six
pep8
pillow
numpy
scipy
pywavelets
